Update Time_Active_Escalator.c
C program that calculates the total number of seconds an escalator was active. Takes input of the number of people and their arrival times. If a person arrives while the escalator is still active from the previous person, it accounts for this overlap to not count the same time period twice.